+++ zhanmao/classes/onlineRent/PayOrderTableViewController.m
@@ -137,7 +137,7 @@
 -(void)payWithType:(PayMethodType)type
 {
     [MBProgressHUD showProgressMessage:@"正在准备支付"];
-    [RentHttpTool getPayOrderStringWithToken:[UserModel token] orderType:[NSString stringWithFormat:@"%ld",(long)self.orderType] payType:[NSString stringWithFormat:@"%d",(int)type] orderId:self.orderModel.idd success:^(NSDictionary *dictionary) {
+    [RentHttpTool getPayOrderStringWithToken:[UserModel token] orderType:[NSString stringWithFormat:@"%ld",(long)self.orderType] payType:[NSString stringWithFormat:@"%ld",(long)type] orderId:self.orderModel.idd success:^(NSDictionary *dictionary) {
         NSLog(@"%@",dictionary);
         [MBProgressHUD hide];
         NSString* msg=[dictionary valueForKey:@"message"];
